CandidateCardComponent fixes
