+++ projects/lib-ui-components/src/lib/lib-ui-components.models.ts
@@ -1,0 +1,15 @@
+
+export interface CandidateInfo {
+  id: string;
+  firstName: string;
+  lastName: string;
+  fullName: string;
+  description?: string;
+  imageUrl?: string;
+  website?: string;
+}
+
+export interface CommitteeData {
+  raised?: number;
+  donors?: number;
+}
